[IMP] website_forum: remove variable by changing the qweb loop variable
before this commit, as the qweb for loop variable is
overriding the value passed from the controller and
causing the bug, which displays wrong forum user
image and verification email warning is fixed by [1]
adding temporary variable before the qweb loop
and restoring after the qweb loop.

after this commit, the temporary variable will be
removed by adjusting the qweb loop variable.

1. https://github.com/odoo/odoo/commit/5f2a0bea54a4daca565d3ba3be32ce5582253cca